feat: Return most relevant tag when multiple tags exist on one commit
Previously, when there were two tags on one commit, neither would be used.
This change that by iterating over tag matches until one matching semver is
found, and uses that one. If no match is found, it falls back to the previous
logic that has been in place to preserve compatibility.